+++ certbot/_internal/display/__init__.py
@@ -1,0 +1,1 @@
+"""Certbot display utilities."""
