+++ certbot/display/util.py
@@ -10,7 +10,7 @@ from certbot import errors
 from certbot import interfaces
 from certbot.compat import misc
 from certbot.compat import os
-from certbot.display import completer
+from certbot._internal.display import completer
 
 logger = logging.getLogger(__name__)
